botoes de upload e relatorio para dentro do componente
